gdk: move event related functions to Event
this also adds some missing doc aliases
